move aws client fixtures to conftest.py
